Create repo for aws directory for DDNS website
Linux Apache MariaDB on my LAN accessed by IP Address site for the
lam1ak.asuscomm.com Dynamic Domain Name Service subdomain
possibly being served by the instance.

The gci/html/index.html file is manually updated when the LAM LAN WAN address
changes.

The site references and depnds on some /var/www/html/Public/Scripts
and /Public defined in other apache2 configuration.

This set of files has a long history and used to be clones of files in the
web space at GCI that used to be optionally provided to every ISP user along
with ftp access.

Within that set of files is IPStatus.html that used to be dynamically
generated hourly and uploaded to the Static Web Space.

This allowed me to know the new IP address if it changed when I was not home.
I could then access my site and use the data to manually update the DNS A
records for my domains.  With a Dynamic Domain Name Service and DNS CNAME
records this is no longer necessary.

When I would manually update the DNS A records for my domains I would
update the HEADER.html file at the root of this site and the About page
as well.

This site is maintained and updated mostly for nostalgic reasons.

I rearrainged the structure with html content in a subdirectory of the same
name and the apache2 configuration file at the top level so that all is in
one directory before initializing this git repository.

I should have put this under git control a long time ago.

The RCS archives remain and the HEADER.html and About/index.html rlog
and differences record a lot of IP address changes.

 cd /var/www
 mkdir t
 mv gci t
 mv t gci
 cd gci
 mv gci html

 sed 's|DocumentRoot \(.*\)|DocumentRoot "/var/www/gci/html"|' \
 /etc/apache2/sites-available/082_gci.conf > gci_apache2.conf

 sed -i 's|RewriteCond /var/www/gci/%1.html -f|RewriteCond /var/www/gci/html/%1.html -f|' \
 gci_apache2.conf

 scp aws:/var/www/aws/README.md .

 scp aws:/var/www/aws/.gitignore .

 ln -s /var/www/gci/gci_apache2.conf \
 /etc/apache2/sites-available/082_gci.conf

 sudo a2ensite 082_gci
 sudo systemctl reload apache2